chore: update carousel
