+++ src/packages/_utils/index.ts
@@ -4,6 +4,3 @@ export * from './types';
 
 export * from './isType';
 
-export function ResizeObserver(fn) {
-  return fn;
-}
