+++ src/packages/carousel/index.ts
@@ -1,9 +1,10 @@
 import type { App } from 'vue';
-import ECarousel from './src/carousel.vue';
-
+import ECarousel from './src';
+import ECarouselItem from './src/carousel-item.vue';
 ECarousel.install = (app: App) => {
   app.component(ECarousel.name, ECarousel);
+  app.component(ECarouselItem.name, ECarouselItem);
 };
 
-export { ECarousel };
+export { ECarousel, ECarouselItem };
 export default ECarousel;
