--- src/packages/carousel/src/carousel.scss
@@ -1,3 +1,0 @@
-.ra-carousel__carousel-item { left: 0; top: 0; width: 100%; height: 100%; position: absolute; display: flex; justify-content: center; align-items: center; }
-
-.ra-carousel__carousel-item.is-animating { transition: transform 0.4s ease-in-out; }
