+++ src/packages/components.ts
@@ -14,3 +14,5 @@ export * from './upload';
 export * from './input-number';
 export * from './date-picker';
 export * from './avatar';
+
+export * from './carousel';
